+++ agave/collections/redis/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['RedisCollection']
+
+from .redis_collection import RedisCollection
