--- agave/repositories/__init__.py
@@ -1,14 +1,0 @@
-__all__ = [
-    'BaseRepository',
-    'MongoRepository',
-    'BaseModel',
-    'String',
-    'RedisRepository',
-]
-
-from agave.repositories.redis.base_redis import BaseModel
-from agave.repositories.redis.helpers_redis import String
-from agave.repositories.redis.redis_repository import RedisRepository
-
-from .base_repository import BaseRepository
-from .mongodb_repository import MongoRepository
