+++ examples/chalicelib/models/models_redis/accounts_redis.py
@@ -2,8 +2,9 @@ import datetime as dt
 
 from rom import DateTime, util
 
+from agave.collections.redis.base_redis import BaseModel
+from agave.collections.redis.helpers_redis import String
 from agave.models.helpers import uuid_field
-from agave.repositories import BaseModel, String
 
 DEFAULT_MISSING_DATE = dt.datetime.utcfromtimestamp(0)
