+++ examples/chalicelib/models/models_redis/transactions.py
@@ -1,7 +1,8 @@
 from rom import Float, util
 
+from agave.collections.redis.base_redis import BaseModel
+from agave.collections.redis.helpers_redis import String
 from agave.models.helpers import uuid_field
-from agave.repositories import BaseModel, String
 
 
 class TransactionRedis(BaseModel):
