+++ examples/chalicelib/resources/resources_redis/transactions_redis.py
@@ -1,5 +1,5 @@
-from agave.filters import generic_query_redis
-from agave.repositories import RedisRepository
+from agave.collections.redis import RedisCollection
+from agave.collections.redis.filters_redis import generic_query_redis
 from examples.chalicelib.models import TransactionRedis as Model
 from examples.chalicelib.validators import TransactionQuery
 
@@ -8,6 +8,6 @@ from ..base import app
 
 @app.resource('/transactions_redis')
 class TransactionRedis:
-    repository = RedisRepository(Model, generic_query_redis)
+    collection = RedisCollection(Model, generic_query_redis)
     query_validator = TransactionQuery
     get_query_filter = generic_query_redis
